+++ src/types/rx-error.d.ts
@@ -3,21 +3,24 @@ import {
     RxSchema
 } from '../rx-schema';
 import { RxPlugin } from './rx-plugin';
+import { ERROR_MESSAGES } from '../plugins/dev-mode/error-messages';
+
+type KeyOf<T extends object> = Extract<keyof T, string>;
+export type RxErrorKey = KeyOf<typeof ERROR_MESSAGES>;
 
 export declare class RxError extends Error {
     readonly rxdb: boolean; // always true, use this to detect if its an rxdb-error
     readonly parameters: RxErrorParameters; // an object with parameters to use the programatically
-    readonly code: string; // error-code
+    readonly code: RxErrorKey; // error-code
     readonly typeError: false; // true if is TypeError
 }
 
 export declare class RxTypeError extends TypeError {
     readonly rxdb: boolean; // always true, use this to detect if its an rxdb-error
     readonly parameters: RxErrorParameters; // an object with parameters to use the programatically
-    readonly code: string; // error-code
+    readonly code: RxErrorKey; // error-code
     readonly typeError: true; // true if is TypeError
 }
-
 
 /**
  * this lists all possible parameters